[EI-200] [web] refactor: Refactor import paths and file names for indicator list components
